requirements.txt
# Data Cleaning & Preprocessing
pandas==2.2.2
numpy==1.26.4
spacy==3.8.4
nltk==3.9.1
textblob==0.19.0

# Natural Language Processing (NLP)
wordcloud==1.9.4

# Database Interaction
sqlalchemy==2.0.34
mysql-connector-python==9.2.0

# Utilities
json
string